FaceRecognition_V3.0/requirement.txt
numpy
keras
tensorflow-gpu
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
threading

